+++ src/app/feature/my-patients/my-patients.component.ts
@@ -10,12 +10,14 @@ import { Observable, Subscription } from 'rxjs';
 import { User } from '../../../app/core/auth/user.model';
 import { FileService } from '../../../app/shared/services/file.service';
 import { take } from 'rxjs/operators';
+import { Router } from '@angular/router';
+import { RegistryService } from '../registry/registry.service';
 
 @Component({
   selector: 'app-my-patients',
   templateUrl: './my-patients.component.html',
   styleUrls: ['./my-patients.component.scss'],
-  providers: [MyPatientsService]
+  providers: [RegistryService, MyPatientsService]
 })
 export class MyPatientsComponent implements OnInit, OnDestroy {
   user$: Observable<User>;
@@ -26,9 +28,11 @@ export class MyPatientsComponent implements OnInit, OnDestroy {
   myCathPciData: RegistryModel[];
 
   constructor(
+    private registryService: RegistryService,
     private myPatientsService: MyPatientsService,
     private store: Store<fromRoot.State>,
-    private fileService: FileService
+    private fileService: FileService,
+    private router: Router
   ) {}
 
   async ngOnInit() {
@@ -54,11 +58,29 @@ export class MyPatientsComponent implements OnInit, OnDestroy {
     // this.userSubscription.unsubscribe();
   }
 
+  clickACSx(registryId: string) {
+    this.store.dispatch(new UI.StartLoading());
+    setTimeout(() => {
+      this.router.navigate(['registry/acsx290', registryId]);
+    }, 300);
+  }
+
+  clickCathPci(registryId: string) {
+    this.store.dispatch(new UI.StartLoading());
+    setTimeout(() => {
+      this.router.navigate(['registry/cath-pci50', registryId]);
+    }, 300);
+  }
+
   async exportACSx() {
     const data = await this.myPatientsService.loadMyACSx290sForExport(this.user.staff.staffId);
-    this.fileService.saveJSONtoCSV(data, 'my-registry.csv');
+    this.fileService.saveJSONtoCSV(data, 'my-acsx.csv');
     console.log('export acsx ' + data.length + ' records');
   }
 
-  async exportCathPci() {}
+  async exportCathPci() {
+    const data = await this.myPatientsService.loadMyCathPci50sForExport(this.user.staff.staffId);
+    this.registryService.exportCathPci50AsExcelFile(data, 'my-cathpci');
+    console.log('export cathpci ' + data.length + ' records');
+  }
 }