+++ src/app/feature/my-patients/my-patients.service.ts
@@ -6,6 +6,7 @@ import { map, take } from 'rxjs/operators';
 import { RegistryModel } from '../registry/registry.model';
 import { AuthService } from '../../../app/core/auth/auth.service';
 import { ACSx290Model } from '../registry/acsx290/acsx290.model';
+import { CathPci50Model } from '../registry/cath-pci50/cath-pci50.model';
 
 const DB_REGISTRY = 'Registry';
 
@@ -59,6 +60,49 @@ export class MyPatientsService {
             delete d.sectionB['DOB'];
             delete d.sectionB['SSN'];
             delete d.sectionB['PatAddr'];
+            // tslint:enable: no-string-literal
+
+            d.detail.createdAt =
+              d.detail.createdAt !== null
+                ? (d.detail.createdAt as firebase.firestore.Timestamp).toDate().toISOString()
+                : null;
+            d.detail.modifiedAt =
+              d.detail.modifiedAt !== null
+                ? (d.detail.modifiedAt as firebase.firestore.Timestamp).toDate().toISOString()
+                : null;
+            d.detail.deletedAt =
+              d.detail.deletedAt !== null
+                ? (d.detail.deletedAt as firebase.firestore.Timestamp).toDate().toISOString()
+                : null;
+
+            return d;
+          })
+        ),
+        take(1)
+      )
+      .toPromise();
+  }
+
+  public async loadMyCathPci50sForExport(staffId: string): Promise<CathPci50Model[]> {
+    const cathpci = await this.authService.getAvailableCathPci50s(staffId);
+
+    return this.db
+      .collection<CathPci50Model>('CathPci50')
+      .valueChanges()
+      .pipe(
+        // tslint:disable-next-line: no-string-literal
+        map(data => data.filter(a => cathpci.includes(a.sectionA['registryId']))),
+        map(data =>
+          data.map(d => {
+            // tslint:disable: no-string-literal
+            delete d.sectionA['HN'];
+            delete d.sectionA['AN'];
+            delete d.sectionA['LastName'];
+            delete d.sectionA['FirstName'];
+            delete d.sectionA['MidName'];
+            delete d.sectionA['DOB'];
+            delete d.sectionA['SSN'];
+            delete d.sectionA['ZipCode'];
             // tslint:enable: no-string-literal
 
             d.detail.createdAt =