+++ src/app/feature/registry/cath-pci50-list/cath-pci50-list.component.ts
@@ -62,6 +62,13 @@ export class CathPci50ListComponent implements OnInit, OnDestroy {
     this.userSubscription.unsubscribe();
   }
 
+  click(registryId: string) {
+    this.store.dispatch(new UI.StartLoading());
+    setTimeout(() => {
+      this.router.navigate(['registry/cath-pci50', registryId]);
+    }, 300);
+  }
+
   create() {
     this.store.dispatch(new UI.StartLoading());
     setTimeout(() => {